Crate tikv

TiKV - A distributed key/value database

TiKV ("Ti" stands for Titanium) is an open source distributed transactional key-value database. Unlike other traditional NoSQL systems, TiKV not only provides classical key-value APIs, but also transactional APIs with ACID compliance. TiKV was originally created to complement TiDB, a distributed HTAP database compatible with the MySQL protocol.

The design of TiKV is inspired by some great distributed systems from Google, such as BigTable, Spanner, and Percolator, and some of the latest achievements in academia in recent years, such as the Raft consensus algorithm.

Modules

config
Configuration for the entire server.

coprocessor
Handles simple SQL query executors locally.

import
This mod contains components to support rapid data import with the project tidb-lightning.

read_pool

server

storage
This module contains TiKV's transaction layer. It lowers high-level, transactional commands to low-level (raw key-value) interactions with persistent storage.

Functions

log_tikv_info
Prints the tikv version information to the standard output.

tikv_version_info
Returns the tikv version information.