Module tikv::config

Configuration for the entire server.

TiKV is configured through the TiKvConfig type, which is in turn made up of many other configuration types.

Modules

log_level_serde

Structs

BackupConfig

CdcConfig

ConfigController
ConfigController use to register each module's config manager, and dispatch the change of config to corresponding managers or return the change if the incoming change is invalid.

ConfigInner

CoprReadPoolConfig

DBConfigManger

DbConfig

DefaultCfConfig

LockCfConfig

MetricConfig

RaftCfConfig

RaftDbConfig

RaftDefaultCfConfig

RaftEngineConfig

ReadPoolConfig

StorageReadPoolConfig

TIKVCONFIG_TYPED

TiKvConfig

TitanCfConfig

TitanDBConfig

UnifiedReadPoolConfig

VersionCfConfig

WriteCfConfig

Enums

DBType

Module

Constants

DEFAULT_COPROCESSOR_READPOOL_MIN_CONCURRENCY

DEFAULT_READPOOL_MAX_TASKS_PER_WORKER

DEFAULT_READPOOL_STACK_SIZE_MB

DEFAULT_ROCKSDB_SUB_DIR

DEFAULT_STORAGE_READPOOL_MAX_CONCURRENCY

DEFAULT_STORAGE_READPOOL_MIN_CONCURRENCY

LAST_CONFIG_FILE

LOCKCF_MAX_MEM

LOCKCF_MIN_MEM

MAX_BLOCK_SIZE

MIN_READPOOL_STACK_SIZE_MB

RAFT_MAX_MEM

RAFT_MIN_MEM

TMP_CONFIG_FILE

UNIFIED_READPOOL_MIN_CONCURRENCY

Functions

check_critical_config
Prevents launching with an incompatible configuration

config_to_slice

config_value_to_string

get_background_job_limit

get_last_config

memory_mb_for_cf

persist_config
Persists config to last_tikv.toml

to_change_value

to_config_change

to_toml_encode

write_config