Module tikv::coprocessor::interceptors::concurrency_limiter

Structs

ConcurrencyLimiter

__ConcurrencyLimiterProjection

__ConcurrencyLimiterProjectionRef

Enums

LimitationState

Traits

ConcurrencyLimiterMustNotImplDrop

Functions

__pin_project_assert_not_repr_packed_ConcurrencyLimiter

__unpin_scope_ConcurrencyLimiter

limit_concurrency
Limits the concurrency of heavy tasks by limiting the time spent on executing fut before forcing to acquire a semaphore permit.