Struct tikv::coprocessor::tracker::Tracker

pub struct Tracker {
    request_begin_at: Instant,
    item_begin_at: Instant,
    current_stage: TrackerState,
    wait_time: Duration,
    schedule_wait_time: Duration,
    snapshot_wait_time: Duration,
    handler_build_time: Duration,
    req_lifetime: Duration,
    item_process_time: Duration,
    total_process_time: Duration,
    total_storage_stats: Statistics,
    total_perf_stats: PerfStatisticsDelta,
    pub req_ctx: ReqContext,
}

Track coprocessor requests to update statistics and provide slow logs.

Fields
request_begin_at: Instantitem_begin_at: Instantcurrent_stage: TrackerStatewait_time: Durationschedule_wait_time: Durationsnapshot_wait_time: Durationhandler_build_time: Durationreq_lifetime: Durationitem_process_time: Durationtotal_process_time: Durationtotal_storage_stats: Statisticstotal_perf_stats: PerfStatisticsDeltareq_ctx: ReqContext
Implementations

impl Tracker

pub fn new(req_ctx: ReqContext) -> Tracker

Initialize the tracker. Normally it is called outside future pool's factory context, because the future pool might be full and we need to wait it. This kind of wait time has to be recorded.

pub fn on_scheduled(&mut self)

pub fn on_snapshot_finished(&mut self)

pub fn on_begin_all_items(&mut self)

pub fn on_begin_item(&mut self)

pub fn on_finish_item(
&mut self,
some_storage_stats: Option<Statistics>,
perf_statistics: PerfStatisticsDelta
)

pub fn collect_storage_statistics(&mut self, storage_stats: Statistics)

pub fn get_item_exec_details(&self) -> ExecDetails

Get current item's ExecDetail according to previous collected metrics. TiDB asks for ExecDetail to be printed in its log. WARN: TRY BEST NOT TO USE THIS FUNCTION.

pub fn get_exec_details(&self) -> ExecDetails

Get ExecDetail according to previous collected metrics. TiDB asks for ExecDetail to be printed in its log.

fn exec_details(&self, measure: Duration) -> ExecDetails

pub fn on_finish_all_items(&mut self)

fn track(&mut self)