Trait tikv::coprocessor::RequestHandler

pub trait RequestHandler: Send {
    fn handle_request<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<Response>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn handle_streaming_request(&mut self) -> Result<(Option<Response>, bool)> { ... }

    fn collect_scan_statistics(&mut self, _dest: &mut Statistics) { ... }

    fn into_boxed(self) -> Box<dyn RequestHandler>
    where
        Self: 'static + Sized,
    { ... }
}

An interface for all kind of Coprocessor request handlers.

Provided methods

fn handle_request<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<Response>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Processes current request and produces a response.

fn handle_streaming_request(&mut self) -> Result<(Option<Response>, bool)>

Processes current request and produces streaming responses.

fn collect_scan_statistics(&mut self, _dest: &mut Statistics)

Collects scan statistics generated in this request handler so far.

fn into_boxed(self) -> Box<dyn RequestHandler> where
Self: 'static + Sized,

Loading content...