Struct tikv::server::debug::Debugger

pub struct Debugger<ER: RaftEngine> {
    engines: Engines<RocksEngine, ER>,
    cfg_controller: ConfigController,
}

Fields
engines: Engines<RocksEngine, ER>cfg_controller: ConfigController
Implementations

impl<ER: RaftEngine> Debugger<ER>

pub fn new(
engines: Engines<RocksEngine, ER>,
cfg_controller: ConfigController
) -> Debugger<ER>

pub fn get_engine(&self) -> &Engines<RocksEngine, ER>

pub fn get_all_meta_regions(&self) -> Result<Vec<u64>>

Get all regions holding region meta data from raft CF in KV storage.

fn get_db_from_type(&self, db: DBType) -> Result<&Arc<DB>>

pub fn get(&self, db: DBType, cf: &str, key: &[u8]) -> Result<Vec<u8>>

pub fn raft_log(&self, region_id: u64, log_index: u64) -> Result<Entry>

pub fn region_info(&self, region_id: u64) -> Result<RegionInfo>

pub fn region_size<T: AsRef<str>>(
&self,
region_id: u64,
cfs: Vec<T>
) -> Result<Vec<(T, usize)>>

pub fn scan_mvcc(
&self,
start: &[u8],
end: &[u8],
limit: u64
) -> Result<MvccInfoIterator<RocksEngineIterator>>

Scan MVCC Infos for given range [start, end).

pub fn raw_scan(
&self,
start: &[u8],
end: &[u8],
limit: usize,
cf: &str
) -> Result<Vec<(Vec<u8>, Vec<u8>)>>

Scan raw keys for given range [start, end) in given cf.

pub fn compact(
&self,
db: DBType,
cf: &str,
start: &[u8],
end: &[u8],
threads: u32,
bottommost: BottommostLevelCompaction
) -> Result<()>

Compact the cf[start..end) in the db.

pub fn set_region_tombstone(
&self,
regions: Vec<Region>
) -> Result<Vec<(u64, Error)>>

Set regions to tombstone by manual, and apply other status(such as peers, version, and key range) from region which comes from PD normally.

pub fn set_region_tombstone_by_id(
&self,
regions: Vec<u64>
) -> Result<Vec<(u64, Error)>>

pub fn recover_regions(
&self,
regions: Vec<Region>,
read_only: bool
) -> Result<Vec<(u64, Error)>>

pub fn recover_all(&self, threads: usize, read_only: bool) -> Result<()>

pub fn bad_regions(&self) -> Result<Vec<(u64, Error)>>

pub fn remove_failed_stores(
&self,
store_ids: Vec<u64>,
region_ids: Option<Vec<u64>>
) -> Result<()>

pub fn recreate_region(&self, region: Region) -> Result<()>

pub fn get_store_id(&self) -> Result<u64>

pub fn get_cluster_id(&self) -> Result<u64>

pub fn modify_tikv_config(
&self,
config_name: &str,
config_value: &str
) -> Result<()>

fn get_region_state(&self, region_id: u64) -> Result<RegionLocalState>

pub fn get_region_properties(
&self,
region_id: u64
) -> Result<Vec<(String, String)>>

pub fn get_range_properties(
&self,
start: &[u8],
end: &[u8]
) -> Result<Vec<(String, String)>>