Module tikv::server::gc_worker::gc_manager

Structs

AutoGcConfig
The configurations of automatic GC.

GcManager
Controls how GC runs automatically on the TiKV. It polls safe point periodically, and when the safe point is updated, GcManager will start to scan all regions (whose leader is on this TiKV), and does GC on all those regions.

GcManagerContext
Used to check if GcManager should be stopped.

GcManagerHandle
Wraps JoinHandle of GcManager and helps to stop the GcManager synchronously.

Enums

GcManagerError
The only error that will break GcManager's process is that the GcManager is interrupted by others, maybe due to TiKV shutting down.

GcManagerState
Used to represent the state of GcManager.

Constants

BEGIN_KEY

POLL_SAFE_POINT_INTERVAL_SECS

PROCESS_TYPE_GC

PROCESS_TYPE_SCAN

Functions

set_status_metrics

Type Definitions

GcManagerResult