Module tikv::server::gc_worker::gc_worker

Structs

GcRunner
Used to perform GC operations on the engine.

GcWorker
Used to schedule GC operations.

Enums

GcTask

Constants

GC_LOG_DELETED_VERSION_THRESHOLD
After the GC delete versions of a key, output a message to the log if at least this many versions are deleted.

GC_LOG_FOUND_VERSION_THRESHOLD
After the GC scan of a key, output a message to the log if there are at least this many versions of the key.

GC_MAX_EXECUTING_TASKS

GC_TASK_SLOW_SECONDS

Traits

GcSafePointProvider
Provides safe point.

Functions

handle_gc_task_schedule_error
When we failed to schedule a GcTask to GcRunner, use this to handle the ScheduleError.

schedule_gc
Schedules a GcTask to the GcRunner.

sync_gc
Does GC synchronously.