Struct tikv::server::lock_manager::deadlock::Detector

pub struct Detector<S, P> where
    S: StoreAddrResolver + 'static,
    P: PdClient + 'static,  {
    store_id: u64,
    env: Arc<Environment>,
    leader_info: Option<(u64, String)>,
    leader_client: Option<Client>,
    pd_client: Arc<P>,
    resolver: S,
    security_mgr: Arc<SecurityManager>,
    waiter_mgr_scheduler: WaiterMgrScheduler,
    inner: Rc<RefCell<Inner>>,
}

Detector is used to detect deadlocks between transactions. There is a leader in the cluster which collects all wait_for_entry from other followers.

Fields
store_id: u64

The store id of the node.

env: Arc<Environment>

Used to create clients to the leader.

leader_info: Option<(u64, String)>

The leader's id and address if exists.

leader_client: Option<Client>

The connection to the leader.

pd_client: Arc<P>

Used to get the leader of leader region from PD.

resolver: S

Used to resolve store address.

security_mgr: Arc<SecurityManager>

Used to connect other nodes.

waiter_mgr_scheduler: WaiterMgrScheduler

Used to schedule Deadlock msgs to the waiter manager.

inner: Rc<RefCell<Inner>>
Implementations

impl<S, P> Detector<S, P> where
S: StoreAddrResolver + 'static,
P: PdClient + 'static,

pub fn new(
store_id: u64,
pd_client: Arc<P>,
resolver: S,
security_mgr: Arc<SecurityManager>,
waiter_mgr_scheduler: WaiterMgrScheduler,
cfg: &Config
) -> Self

fn is_leader(&self) -> bool

Returns true if it is the leader of the deadlock detector.

fn reset(&mut self, role: Role)

Resets to the initial state.

fn refresh_leader_info(&mut self) -> bool

Refreshes the leader info. Returns true if the leader exists.

fn get_leader_info(&self) -> Result<Option<(u64, String)>>

Gets leader info from PD.

fn resolve_store_address(&self, store_id: u64) -> Result<String>

Resolves store address.

fn update_leader_info(&mut self, leader_id: u64, leader_addr: String)

Updates the leader info.

fn change_role(&mut self, role: Role)

Resets state if role changes.

fn reconnect_leader(&mut self, handle: &Handle)

Reconnects the leader. The leader info must exist.

fn send_request_to_leader(
&mut self,
handle: &Handle,
tp: DetectType,
txn_ts: TimeStamp,
lock: Lock
) -> bool

Returns true if sends successfully.

If the client is None, reconnects the leader first, then sends the request to the leader. If sends failed, sets the client to None for retry.

fn handle_detect_locally(&self, tp: DetectType, txn_ts: TimeStamp, lock: Lock)

fn handle_detect(
&mut self,
handle: &Handle,
tp: DetectType,
txn_ts: TimeStamp,
lock: Lock
)

Handles detect requests of itself.

fn handle_detect_rpc(
&self,
handle: &Handle,
stream: RequestStream<DeadlockRequest>,
sink: DuplexSink<DeadlockResponse>
)

Handles detect requests of other nodes.

fn handle_change_role(&mut self, role: Role)

fn handle_change_ttl(&mut self, ttl: Duration)