Struct tikv::server::lock_manager::waiter_manager::Delay

struct Delay {
    inner: Rc<RefCell<DelayInner>>,
    deadline: Instant,
}

Delay is a wrapper of tokio_timer::Delay which has a resolution of one millisecond. It has some extra features than tokio_timer::Delay used by WaiterManager.

Delay performs no work and completes with true once the specified deadline has been reached. If it has been cancelled, it will complete with false at arbitrary time.

Fields
inner: Rc<RefCell<DelayInner>>deadline: Instant
Implementations

impl Delay

fn new(deadline: Instant) -> Self

Create a new Delay instance that elapses at deadline.

fn reset(&self, deadline: Instant)

Resets the instance to an earlier deadline.

fn cancel(&self)

Cancels the instance. It will complete with false at arbitrary time.

fn is_cancelled(&self) -> bool