Module tikv::server::raft_client

Structs

BatchMessageBuffer
A buffer for BatchRaftMessage.

CachedQueue
Queue in cache.

ConnectionBuilder

ConnectionPool
A global connection pool.

MessageBuffer
A buffer for non-batch RaftMessage.

Queue
A quick queue for sending raft messages.

RaftCall
Struct tracks the lifetime of a raft or batch_raft RPC.

RaftClient
A raft client that can manages connections correctly.

SnapshotReporter
Reporter reports whether a snapshot is sent successfully.

StreamBackEnd
StreamBackEnd watches lifetime of a connection and handles reconnecting, spawn new RPC.

Constants

GRPC_SEND_MSG_BUF

QUEUE_CAPACITY

RAFT_MSG_MAX_BATCH_SIZE

_ON_RESOLVE_FP

Statics

CONN_ID

Traits

Buffer

Functions

grpc_error_is_unimplemented

maybe_backoff

report_unreachable

start
A future that drives the life cycle of a connection.