Function tikv::server::raft_client::start

async fn start<S, R>(
    back_end: StreamBackEnd<S, R>, 
    conn_id: usize, 
    pool: Arc<Mutex<ConnectionPool>>
) where
    S: StoreAddrResolver + Send,
    R: RaftStoreRouter<RocksEngine> + Unpin + Send + 'static, 

A future that drives the life cycle of a connection.

The general progress of connection is:

resolve address

connect

make batch call

fallback to legacy API if incompatible

Every failure during the process should trigger retry automatically.