Module tikv::server::raft_client::fresh

Structs

BatchMessageBuffer
A buffer for BatchRaftMessage.

MessageBuffer
A buffer for non-batch RaftMessage.

Queue
A quick queue for sending raft messages.

RaftCall
Struct tracks the lifetime of a raft or batch_raft RPC.

SnapshotReporter
Reporter reports whether a snapshot is sent successfully.

StreamBackEnd
StreamBackEnd watches lifetime of a connection and handles reconnecting, spawn new RPC.

Constants

GRPC_SEND_MSG_BUF

RAFT_MSG_MAX_BATCH_SIZE

Statics

CONN_ID

Traits

Buffer

Functions

start
A future that drives the life cycle of a connection.