Function tikv::server::raft_client::fresh::start

async fn start<S, R>(back_end: StreamBackEnd<S, R>) where
    S: StoreAddrResolver,
    R: RaftStoreRouter<RocksEngine> + 'static, 

A future that drives the life cycle of a connection.

The general progress of connection is:

1. resolve address
2. connect
3. make batch call
4. fallback to legacy API if incompatible

Every failure during the process should trigger retry automatically.