Module tikv::server::service::diagnostics::log

Structs

LogIterator

NUM_REGEX

Enums

Error

Constants

LOG_ITEM_BATCH_SIZE

TIMESTAMP_LENGTH

Functions

batch_log_item

is_log_file

parse
Parses the single log line and retrieve the log meta and log body.

parse_end_time

parse_level

parse_start_time

parse_time

parse_time_range
Parses the start time and end time of a log file and return the maximal and minimal timestamp in unix milliseconds.

search