Module tikv::storage

Interact with persistent storage.

The Storage structure provides raw and transactional APIs on top of a lower-level Engine.

There are multiple Engine implementations, RaftKv is used by the Server. The BTreeEngine and RocksEngine are used for testing only.

Re-exports

pub use self::errors::get_error_kind_from_header;

pub use self::errors::get_tag_from_header;

pub use self::errors::Error;

pub use self::errors::ErrorHeaderKind;

pub use self::errors::ErrorInner;

pub use self::kv::CbContext;

pub use self::kv::Engine;

pub use self::kv::Iterator;

pub use self::kv::ScanMode;

pub use self::kv::Snapshot;

pub use self::txn::ProcessResult;

Modules

config
Storage configuration.

errors

kv

lock_manager

metrics

mvcc
Multi-version concurrency control functionality.

read_pool

test_util

txn
Storage Transactions

types
Core data types.

Structs

CfStatistics
Statistics collects the ops taken when fetching data.

Cursor

FlowStatistics

PrewriteResult

RocksEngine
The RocksEngine is based on RocksDB.

SnapshotStore

Statistics

Storage
Storage implements transactional KV APIs and raw KV APIs on a given Engine. An Engine provides low level KV functionality. Engine has multiple implementations. When a TiKV server is running, a RaftKv will be the underlying Engine of Storage. The other two types of engines are for test purpose.

TestEngineBuilder
A builder to build a temporary RocksEngine.

TestStorageBuilder
A builder to build a temporary Storage<E>.

Enums

PessimisticLockRes

SecondaryLocksStatus

StorageCallback

TxnStatus
Represents the status of a transaction.

Traits

FlowStatsReporter

Scanner
Scanners allow retrieving items or batches from a scan result.

Store

Functions

async_commit_check_keys

build_read_pool

build_read_pool_for_test

get_priority_tag

Type Definitions

Callback

Result