Struct tikv::storage::kv::btree_engine::BTreeEngineIterator

pub struct BTreeEngineIterator {
    tree: Arc<RwLock<BTreeMap<Key, Value>>>,
    cur_key: Option<Key>,
    cur_value: Option<Value>,
    valid: bool,
    bounds: (Bound<Key>, Bound<Key>),
}

Fields
tree: Arc<RwLock<BTreeMap<Key, Value>>>cur_key: Option<Key>cur_value: Option<Value>valid: boolbounds: (Bound<Key>, Bound<Key>)
Implementations

impl BTreeEngineIterator

pub fn new(
tree: Arc<RwLock<BTreeMap<Key, Value>>>,
iter_opt: IterOptions
) -> BTreeEngineIterator

fn seek_to_range_endpoint(
&mut self,
range: (Bound<Key>, Bound<Key>),
left: bool
) -> bool

In general, there are 2 endpoints in a range, the left one and the right one. This method will seek to the left one if left is true, else seek to the right one. Returns true when the endpoint is valid, which means the endpoint exist and in self.bounds.