Trait tikv::storage::kv::Engine

pub trait Engine: Send + Clone + 'static {
    type Snap: Snapshot;
    type Local: LocalEngine;
    fn kv_engine(&self) -> Self::Local;

    fn snapshot_on_kv_engine(
        &self, 
        start_key: &[u8], 
        end_key: &[u8]
    ) -> Result<Self::Snap>;

    fn modify_on_kv_engine(&self, modifies: Vec<Modify>) -> Result<()>;

    fn async_snapshot(
        &self, 
        ctx: SnapContext<'_>, 
        cb: Callback<Self::Snap>
    ) -> Result<()>;

    fn async_write(
        &self, 
        ctx: &Context, 
        batch: WriteData, 
        write_cb: Callback<()>
    ) -> Result<()>;

    fn async_write_ext(
        &self, 
        ctx: &Context, 
        batch: WriteData, 
        write_cb: Callback<()>, 
        _proposed_cb: Option<ExtCallback>, 
        _committed_cb: Option<ExtCallback>
    ) -> Result<()> { ... }

    fn write(&self, ctx: &Context, batch: WriteData) -> Result<()> { ... }

    fn release_snapshot(&self) { ... }

    fn snapshot(&self, ctx: SnapContext<'_>) -> Result<Self::Snap> { ... }

    fn put(&self, ctx: &Context, key: Key, value: Value) -> Result<()> { ... }

    fn put_cf(
        &self, 
        ctx: &Context, 
        cf: CfName, 
        key: Key, 
        value: Value
    ) -> Result<()> { ... }

    fn delete(&self, ctx: &Context, key: Key) -> Result<()> { ... }

    fn delete_cf(&self, ctx: &Context, cf: CfName, key: Key) -> Result<()> { ... }

    fn get_properties(
        &self, 
        start: &[u8], 
        end: &[u8]
    ) -> Result<RocksTablePropertiesCollection> { ... }

    fn get_properties_cf(
        &self, 
        _: CfName, 
        _start: &[u8], 
        _end: &[u8]
    ) -> Result<RocksTablePropertiesCollection> { ... }

    fn get_mvcc_properties_cf(
        &self, 
        _: CfName, 
        _safe_point: TimeStamp, 
        _start: &[u8], 
        _end: &[u8]
    ) -> Option<MvccProperties> { ... }
}

Engine defines the common behaviour for a storage engine type.

Associated Types

type Snap: Snapshot

type Local: LocalEngine

Loading content...
Required methods

fn kv_engine(&self) -> Self::Local

Local storage engine.

fn snapshot_on_kv_engine(
&self,
start_key: &[u8],
end_key: &[u8]
) -> Result<Self::Snap>

fn modify_on_kv_engine(&self, modifies: Vec<Modify>) -> Result<()>

Write modifications into internal local engine directly.

fn async_snapshot(
&self,
ctx: SnapContext<'_>,
cb: Callback<Self::Snap>
) -> Result<()>

fn async_write(
&self,
ctx: &Context,
batch: WriteData,
write_cb: Callback<()>
) -> Result<()>

Loading content...
Provided methods

fn async_write_ext(
&self,
ctx: &Context,
batch: WriteData,
write_cb: Callback<()>,
_proposed_cb: Option<ExtCallback>,
_committed_cb: Option<ExtCallback>
) -> Result<()>

Writes data to the engine asynchronously with some extensions.

When the write request is proposed successfully, the proposed_cb is invoked. When the write request is finished, the write_cb is invoked.

fn write(&self, ctx: &Context, batch: WriteData) -> Result<()>

fn release_snapshot(&self)

fn snapshot(&self, ctx: SnapContext<'_>) -> Result<Self::Snap>

fn put(&self, ctx: &Context, key: Key, value: Value) -> Result<()>

fn put_cf(
&self,
ctx: &Context,
cf: CfName,
key: Key,
value: Value
) -> Result<()>

fn delete(&self, ctx: &Context, key: Key) -> Result<()>

fn delete_cf(&self, ctx: &Context, cf: CfName, key: Key) -> Result<()>

fn get_properties(
&self,
start: &[u8],
end: &[u8]
) -> Result<RocksTablePropertiesCollection>

fn get_properties_cf(
&self,
_: CfName,
_start: &[u8],
_end: &[u8]
) -> Result<RocksTablePropertiesCollection>

fn get_mvcc_properties_cf(
&self,
_: CfName,
_safe_point: TimeStamp,
_start: &[u8],
_end: &[u8]
) -> Option<MvccProperties>

Loading content...