Enum tikv::storage::mvcc::TxnCommitRecord

pub enum TxnCommitRecord {
    None {
        overlapped_write: Option<Write>,
    },
    SingleRecord {
        commit_ts: TimeStamp,
        write: Write,
    },
    OverlappedRollback {
        commit_ts: TimeStamp,
    },
}

The result of get_txn_commit_record, which is used to get the status of a specified transaction from write cf.

Variants

None

The commit record of the given transaction is not found. But it's possible that there's another transaction's commit record, whose commit_ts equals to the current transaction's start_ts. That kind of record will be returned via the overlapped_write field. In this case, if the current transaction is to be rolled back, the overlapped_write must not be overwritten.

Fields of None

overlapped_write: Option<Write>

SingleRecord

Found the transaction's write record.

Fields of SingleRecord

commit_ts: TimeStampwrite: Write

OverlappedRollback

The transaction's status is found in another transaction's record's overlapped_rollback field. This may happen when the current transaction's start_ts is the same as the commit_ts of another transaction on this key.

Fields of OverlappedRollback

commit_ts: TimeStamp

Implementations

impl TxnCommitRecord

pub fn exist(&self) -> bool

pub fn info(&self) -> Option<(TimeStamp, WriteType)>

pub fn unwrap_single_record(self) -> (TimeStamp, WriteType)

pub fn unwrap_overlapped_rollback(self) -> TimeStamp

pub fn unwrap_none(self) -> Option<Write>