Module tikv::storage::mvcc::reader::scanner

Re-exports

pub use self::forward::test_util;

pub use self::forward::DeltaScanner;

pub use self::forward::EntryScanner;

Modules

backward

forward

Structs

ScannerBuilder

ScannerConfig

Enums

Scanner

Functions

has_data_in_range

near_load_data_by_write
Reads user key's value in default CF according to the given write CF value (write).

near_reverse_load_data_by_write
Similar to near_load_data_by_write, but accepts a BackwardCursor and use near_seek_for_prev internally.

seek_for_valid_value
Seek for the last written value. The write cursor must indicate a data key of the user key of which ts <= after_ts. Return None if cannot find any valid write record or found a delete record.

seek_for_valid_write
Seek for the next valid (write type == Put or Delete) write record. The write cursor must indicate a data key of the user key of which ts <= after_ts. Return None if cannot find any valid write record.