Module tikv::storage::mvcc::txn

Structs

GcInfo

MvccTxn
An abstraction of a locally-transactional MVCC key-value store

ReleasedLock
ReleasedLock contains the information of the lock released by commit, rollback and so on. It's used by LockManager to wake up transactions waiting for locks.

Enums

MissingLockAction

Constants

MAX_TXN_WRITE_SIZE

Functions

make_rollback
Generate the Write record that should be written that means to to perform a specified rollback operation.

make_txn_error