Module tikv::storage::read_pool

Distinct thread pools to handle read commands having different priority levels.

Structs

FuturePoolTicker

Functions

build_read_pool
Build respective thread pools to handle read commands of different priority levels.

build_read_pool_for_test
Build a thread pool that has default tick behavior for testing.