Struct tikv::storage::Cursor

pub struct Cursor<I: Iterator> {
    iter: I,
    scan_mode: ScanMode,
    min_key: Option<Vec<u8>>,
    max_key: Option<Vec<u8>>,
    cur_key_has_read: Cell<bool>,
    cur_value_has_read: Cell<bool>,
}

Fields
iter: Iscan_mode: ScanModemin_key: Option<Vec<u8>>max_key: Option<Vec<u8>>cur_key_has_read: Cell<bool>cur_value_has_read: Cell<bool>
Implementations

impl<I: Iterator> Cursor<I>

pub fn new(iter: I, mode: ScanMode) -> Self

fn mark_unread(&self)

Mark key and value as unread. It will be invoked once cursor is moved.

fn mark_key_read(&self) -> bool

Mark key as read. Returns whether key was marked as read before this call.

fn mark_value_read(&self) -> bool

Mark value as read. Returns whether value was marked as read before this call.

pub fn seek(&mut self, key: &Key, statistics: &mut CfStatistics) -> Result<bool>

pub fn near_seek(
&mut self,
key: &Key,
statistics: &mut CfStatistics
) -> Result<bool>

Seek the specified key.

This method assume the current position of cursor is around key, otherwise you should use seek instead.

pub fn get(
&mut self,
key: &Key,
statistics: &mut CfStatistics
) -> Result<Option<&[u8]>>

Get the value of specified key.

This method assume the current position of cursor is around key, otherwise you should seek first.

pub fn seek_for_prev(
&mut self,
key: &Key,
statistics: &mut CfStatistics
) -> Result<bool>

pub fn near_seek_for_prev(
&mut self,
key: &Key,
statistics: &mut CfStatistics
) -> Result<bool>

Find the largest key that is not greater than the specific key.

pub fn reverse_seek(
&mut self,
key: &Key,
statistics: &mut CfStatistics
) -> Result<bool>

pub fn near_reverse_seek(
&mut self,
key: &Key,
statistics: &mut CfStatistics
) -> Result<bool>

Reverse seek the specified key.

This method assume the current position of cursor is around key, otherwise you should use reverse_seek instead.

pub fn key(&self, statistics: &mut CfStatistics) -> &[u8]

pub fn value(&self, statistics: &mut CfStatistics) -> &[u8]

pub fn seek_to_first(&mut self, statistics: &mut CfStatistics) -> bool

pub fn seek_to_last(&mut self, statistics: &mut CfStatistics) -> bool

pub fn internal_seek(
&mut self,
key: &Key,
statistics: &mut CfStatistics
) -> Result<bool>

pub fn internal_seek_for_prev(
&mut self,
key: &Key,
statistics: &mut CfStatistics
) -> Result<bool>

pub fn next(&mut self, statistics: &mut CfStatistics) -> bool

pub fn prev(&mut self, statistics: &mut CfStatistics) -> bool

pub fn valid(&self) -> Result<bool>

fn handle_error_status(&self, e: Error) -> Result<()>