Module tikv::storage::txn

Storage Transactions

Re-exports

pub use self::commands::Command;

pub use self::scheduler::Scheduler;

Modules

actions
This file contains the "actions" we perform on a [MvccTxn] and related tests "Actions" here means a group of more basic operations, eg. [MvccTxn::load_lock], [MvccTxn::put_write], which are methods on [MvccTxn], for archiving a certain target

commands
Commands used in the transaction system

latch

sched_pool

scheduler
Scheduler which schedules the execution of storage::Commands.

store

tests

Structs

EntryBatch
A batch of transaction entries.

Error

FixtureStore
A Store that reads on fixtures.

FixtureStoreScanner
A Scanner that scans on fixtures.

SnapshotStore

TransactionProperties

Enums

CommitKind

ErrorInner

ProcessResult
Process result of a command.

TransactionKind

TxnEntry
A transaction entry in underlying storage.

Constants

RESOLVE_LOCK_BATCH_SIZE

Traits

Scanner
Scanners allow retrieving items or batches from a scan result.

Store

TxnEntryScanner
TxnEntryScanner allows retrieving items or batches from a scan result.

TxnEntryStore

Functions

acquire_pessimistic_lock

cleanup
Cleanup the lock if it's TTL has expired, comparing with current_ts. If current_ts is 0, cleanup the lock without checking TTL. If the lock is the primary lock of a pessimistic transaction, the rollback record is protected from being collapsed.

commit

gc

prewrite
Prewrite a single mutation by creating and storing a lock and value.

Type Definitions

Result