Module tikv::storage::txn::actions

This file contains the "actions" we perform on a crate::storage::mvcc::MvccTxn and related tests. "Actions" here means a group of more basic operations, eg. crate::storage::mvcc::MvccReader::load_lock, crate::storage::mvcc::MvccTxn::put_write, which are methods on crate::storage::mvcc::MvccTxn, for archiving a certain target.

Modules

acquire_pessimistic_lock

check_data_constraint

check_txn_status

cleanup

commit

gc

get_old_value

prewrite

tests
This file contains tests and testing tools which affects multiple actions