Struct tikv::storage::txn::commands::TypedCommand

pub struct TypedCommand<T> {
    pub cmd: Command,
    _pd: PhantomData<T>,
}

A Command with its return type, reified as the generic parameter T.

Incoming grpc requests (like CommitRequest, PrewriteRequest) are converted to this type via a series of transformations. That process is described below using CommitRequest as an example:

A CommitRequest is handled by the future_commit method in kv.rs, where it needs to be transformed to a TypedCommand before being passed to the storage.sched_txn_command method.

The From<CommitRequest> impl for TypedCommand gets chosen, and its generic parameter indicates that the result type for this instance of TypedCommand is going to be TxnStatus - one of the variants of the StorageCallback enum.

In the above from method, the details of the commit request are captured by creating an instance of the struct storage::txn::commands::commit::Command via its new method.

This struct is wrapped in a variant of the enum storage::txn::commands::Command. This enum exists to facilitate generic operations over different commands.

Finally, the Command enum variant for Commit is converted to the TypedCommand using the From<Command> impl for TypedCommand.

For other requests, see the corresponding future_ method, the From trait implementation and so on.

Fields
cmd: Command_pd: PhantomData<T>

Track the type of the command's return value.