Module tikv::storage::txn::latch

Structs

Latch
Latch which is used to serialize accesses to resources hashed to the same slot.

Latches
Latches which are used for concurrency control in the scheduler.

Lock
Lock required for a command.

Constants

WAITING_LIST_MAX_CAPACITY

WAITING_LIST_SHRINK_SIZE