Trait tikv::storage::txn::store::TxnEntryScanner

pub trait TxnEntryScanner: Send {
    fn next_entry(&mut self) -> Result<Option<TxnEntry>>;

    fn take_statistics(&mut self) -> Statistics;

    fn scan_entries(&mut self, batch: &mut EntryBatch) -> Result<()> { ... }
}

TxnEntryScanner allows retrieving items or batches from a scan result.

Commonly they are obtained as a result of a entry_scanner operation.

Required methods

fn next_entry(&mut self) -> Result<Option<TxnEntry>>

fn take_statistics(&mut self) -> Statistics

Take statistics.

Loading content...
Provided methods

fn scan_entries(&mut self, batch: &mut EntryBatch) -> Result<()>

Loading content...