Struct tikv::server::gc_worker::GcWorker

pub struct GcWorker<E, RR> where
    E: Engine,
    RR: RaftStoreRouter<RocksEngine> + 'static,  {
    engine: E,
    raft_store_router: RR,
    config_manager: GcWorkerConfigManager,
    scheduled_tasks: Arc<AtomicUsize>,
    refs: Arc<AtomicUsize>,
    worker: Arc<Mutex<FutureWorker<GcTask>>>,
    worker_scheduler: FutureScheduler<GcTask>,
    applied_lock_collector: Option<Arc<AppliedLockCollector>>,
    gc_manager_handle: Arc<Mutex<Option<GcManagerHandle>>>,
    feature_gate: FeatureGate,
}

Used to schedule GC operations.

Fields
engine: Eraft_store_router: RR

raft_store_router is useful to signal raftstore clean region size informations.

config_manager: GcWorkerConfigManagerscheduled_tasks: Arc<AtomicUsize>

How many requests are scheduled from outside and unfinished.

refs: Arc<AtomicUsize>

How many strong references. The worker will be stopped once there are no more references.

worker: Arc<Mutex<FutureWorker<GcTask>>>worker_scheduler: FutureScheduler<GcTask>applied_lock_collector: Option<Arc<AppliedLockCollector>>gc_manager_handle: Arc<Mutex<Option<GcManagerHandle>>>feature_gate: FeatureGate
Implementations

impl<E, RR> GcWorker<E, RR> where
E: Engine,
RR: RaftStoreRouter<RocksEngine>,

pub fn new(
engine: E,
raft_store_router: RR,
cfg: GcConfig,
feature_gate: FeatureGate
) -> GcWorker<E, RR>

pub fn start_auto_gc<S: GcSafePointProvider, R: RegionInfoProvider + Clone + 'static>(
&self,
cfg: AutoGcConfig<S, R>,
safe_point: Arc<AtomicU64>
) -> Result<()>

pub fn start(&mut self) -> Result<()>

pub fn start_observe_lock_apply(
&mut self,
coprocessor_host: &mut CoprocessorHost<RocksEngine>,
concurrency_manager: ConcurrencyManager
) -> Result<()>

pub fn stop(&self) -> Result<()>

pub fn scheduler(&self) -> FutureScheduler<GcTask>

fn check_is_busy<T: 'static>(
&self,
callback: Callback<T>
) -> Option<Callback<T>>

Check whether GCWorker is busy. If busy, callback will be invoked with an error that indicates GCWorker is busy; otherwise, return a new callback that invokes the original callback as well as decrease the scheduled task counter.

pub fn gc(&self, safe_point: TimeStamp, callback: Callback<()>) -> Result<()>

Only for tests.

pub fn unsafe_destroy_range(
&self,
ctx: Context,
start_key: Key,
end_key: Key,
callback: Callback<()>
) -> Result<()>

Cleans up all keys in a range and quickly free the disk space. The range might span over multiple regions, and the ctx doesn’t indicate region. The request will be done directly on RocksDB, bypassing the Raft layer. User must promise that, after calling destroy_range, the range will never be accessed any more. However, destroy_range is allowed to be called multiple times on an single range.

pub fn get_config_manager(&self) -> GcWorkerConfigManager

pub fn physical_scan_lock(
&self,
ctx: Context,
max_ts: TimeStamp,
start_key: Key,
limit: usize,
callback: Callback<Vec<LockInfo>>
) -> Result<()>

pub fn start_collecting(
&self,
max_ts: TimeStamp,
callback: Box<dyn FnOnce(Result<()>) + Send>
) -> Result<()>

pub fn get_collected_locks(
&self,
max_ts: TimeStamp,
callback: Box<dyn FnOnce(Result<(Vec<LockInfo>, bool)>) + Send>
) -> Result<()>

pub fn stop_collecting(
&self,
max_ts: TimeStamp,
callback: Box<dyn FnOnce(Result<()>) + Send>
) -> Result<()>