Struct tikv::server::lock_manager::LockManager

pub struct LockManager {
    waiter_mgr_worker: Option<FutureWorker<Task>>,
    detector_worker: Option<FutureWorker<Task>>,
    waiter_mgr_scheduler: WaiterMgrScheduler,
    detector_scheduler: DetectorScheduler,
    waiter_count: Arc<AtomicUsize>,
    detected: Arc<[CachePadded<Mutex<HashSet<TimeStamp>>>]>,
    pipelined: Arc<AtomicBool>,
}

LockManager has two components working in two threads:

One is the WaiterManager which manages transactions waiting for locks.

The other one is the Detector which detects deadlocks between transactions.

Fields
waiter_mgr_worker: Option<FutureWorker<Task>>detector_worker: Option<FutureWorker<Task>>waiter_mgr_scheduler: WaiterMgrSchedulerdetector_scheduler: DetectorSchedulerwaiter_count: Arc<AtomicUsize>detected: Arc<[CachePadded<Mutex<HashSet<TimeStamp>>>]>

Record transactions which have sent requests to detect deadlock.

pipelined: Arc<AtomicBool>
Implementations

impl LockManager

pub fn new(pipelined: bool) -> Self

pub fn start<S, P>(
&mut self,
store_id: u64,
pd_client: Arc<P>,
resolver: S,
security_mgr: Arc<SecurityManager>,
cfg: &Config
) -> Result<()> where
S: StoreAddrResolver + 'static,
P: PdClient + 'static,

Starts WaiterManager and Detector.

pub fn stop(&mut self)

Stops WaiterManager and Detector.

fn start_waiter_manager(&mut self, cfg: &Config) -> Result<()>

fn stop_waiter_manager(&mut self)

fn start_deadlock_detector<S, P>(
&mut self,
store_id: u64,
pd_client: Arc<P>,
resolver: S,
security_mgr: Arc<SecurityManager>,
cfg: &Config
) -> Result<()> where
S: StoreAddrResolver + 'static,
P: PdClient + 'static,

fn stop_deadlock_detector(&mut self)

pub fn register_detector_role_change_observer(
&self,
host: &mut CoprocessorHost<RocksEngine>
)

Creates a RoleChangeNotifier of the deadlock detector worker and registers it to the CoprocessorHost to observe the role change events of the leader region.

pub fn deadlock_service(
&self,
security_mgr: Arc<SecurityManager>
) -> DeadlockService

Creates a DeadlockService to handle deadlock detect requests from other nodes.

pub fn config_manager(&self) -> LockManagerConfigManager

pub fn get_pipelined(&self) -> Arc<AtomicBool>

fn add_to_detected(&self, txn_ts: TimeStamp)

fn remove_from_detected(&self, txn_ts: TimeStamp) -> bool