Module tikv::storage::txn

Storage Transactions

Re-exports

pub use self::commands::Command;

pub use self::scheduler::Scheduler;

Modules

commands
Commands used in the transaction system

latch

process

sched_pool

scheduler
Scheduler which schedules the execution of storage::Commands.

store

Structs

EntryBatch
A batch of transaction entries.

Error

FixtureStore
A Store that reads on fixtures.

FixtureStoreScanner
A Scanner that scans on fixtures.

SnapshotStore

Enums

ErrorInner

ProcessResult
Process result of a command.

TxnEntry
A transaction entry in underlying storage.

Constants

RESOLVE_LOCK_BATCH_SIZE

Traits

Scanner
Scanners allow retrieving items or batches from a scan result.

Store

TxnEntryScanner
TxnEntryScanner allows retrieving items or batches from a scan result.

TxnEntryStore

Type Definitions

Result