Module tikv::storage::txn::actions

This file contains the "actions" we perform on a [MvccTxn] and related tests "Actions" here means a group of more basic operations, eg. [MvccTxn::load_lock], [MvccTxn::put_write], which are methods on [MvccTxn], for archiving a certain target

Modules

acquire_pessimistic_lock

check_data_constraint

check_txn_status

cleanup

commit

pessimistic_prewrite

prewrite

shared

tests
This file contains tests and testing tools which affects multiple actions