Trait tikv::storage::txn::commands::prewrite::MutationLock

trait MutationLock {
    fn is_pessimistic_lock(&self) -> bool;

    fn into_mutation(self) -> Mutation;
}

The type of mutation and, optionally, its extra information, differing for the optimistic and pessimistic transaction. For optimistic txns, this is Mutation. For pessimistic txns, this is (Mutation, bool), where the bool indicates whether the mutation takes a pessimistic lock or not.

Required methods

fn is_pessimistic_lock(&self) -> bool

fn into_mutation(self) -> Mutation

Loading content...